requirements.txt
fastapi==0.119.0
uvicorn==0.37.0
python-dotenv==1.1.1
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-oauthlib==1.2.3
pydantic==2.12.2
email-validator==2.3.0
requests==2.32.5
